+++ pkg/testing/templates.go
@@ -27,7 +27,6 @@ import (
 	"sigs.k8s.io/yaml"
 
 	"github.com/vmware-tanzu/cartographer/pkg/apis/v1alpha1"
-	"github.com/vmware-tanzu/cartographer/pkg/controllers"
 	"github.com/vmware-tanzu/cartographer/pkg/realizer"
 	"github.com/vmware-tanzu/cartographer/pkg/templates"
 )
@@ -155,120 +154,6 @@ type TTOutputs interface {
 	GetOutputs() (realizer.Outputs, error)
 }
 
-func (i *TemplateTestGivens) getActualObject() (*unstructured.Unstructured, error) {
-	ctx := context.Background()
-
-	workload, err := i.Workload.GetWorkload()
-	if err != nil {
-		return nil, fmt.Errorf("get workload failed: %w", err)
-	}
-
-	apiTemplate, err := i.Template.GetTemplate()
-	if err != nil {
-		return nil, fmt.Errorf("get populated template failed: %w", err)
-	}
-
-	if err = (*apiTemplate).ValidateCreate(); err != nil {
-		return nil, fmt.Errorf("template validation failed: %w", err)
-	}
-
-	template, err := templates.NewReaderFromAPI(*apiTemplate)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get cluster template")
-	}
-
-	if template.IsYTTTemplate() {
-		err = ensureYTTAvailable(ctx)
-		if err != nil {
-			return nil, fmt.Errorf("ensure YTT available: %w", err)
-		}
-	}
-
-	if i.actualBlueprintSupplied() {
-		return i.actualBlueprintStamp(ctx, workload, template)
-	}
-
-	return i.mockedBlueprintStamp(ctx, workload, *apiTemplate, template)
-}
-
-func (i *TemplateTestGivens) actualBlueprintStamp(ctx context.Context, workload *v1alpha1.Workload, template templates.Reader) (*unstructured.Unstructured, error) {
-	supplyChain, err := i.SupplyChain.GetSupplyChain(workload)
-	if err != nil {
-		return nil, fmt.Errorf("get supplychain: %w", err)
-	}
-
-	resource, err := i.getTargetResource(realizer.MakeSupplychainOwnerResources(supplyChain))
-	if err != nil {
-		return nil, fmt.Errorf("get target resource: %w", err)
-	}
-
-	templatingContext := realizer.NewContextGenerator(workload, workload.Spec.Params, supplyChain.Spec.Params)
-
-	resourceLabeler := controllers.BuildWorkloadResourceLabeler(workload, supplyChain)
-	labels := resourceLabeler(*resource, template)
-
-	outputs, err := i.TTOutputs.GetOutputs()
-
-	stamper := templates.StamperBuilder(workload, templatingContext.Generate(template, *resource, outputs, labels), labels)
-	actualStampedObject, err := stamper.Stamp(ctx, template.GetResourceTemplate())
-	if err != nil {
-		return nil, fmt.Errorf("could not stamp: %w", err)
-	}
-
-	return actualStampedObject, nil
-}
-
-func (i *TemplateTestGivens) getTargetResource(resources []realizer.OwnerResource) (*realizer.OwnerResource, error) {
-	targetResourceName, err := i.TargetResource.GetTargetResourceName()
-	if err != nil {
-		return nil, fmt.Errorf("get target resource name: %w", err)
-	}
-
-	for _, resource := range resources {
-		if resource.Name == targetResourceName {
-			return &resource, nil
-		}
-	}
-
-	return nil, fmt.Errorf("did not find a supply chain resource with target name: %s", targetResourceName)
-}
-
-func (i *TemplateTestGivens) actualBlueprintSupplied() bool {
-	return i.SupplyChain != nil
-}
-
-func (i *TemplateTestGivens) mockedBlueprintStamp(ctx context.Context, workload *v1alpha1.Workload, apiTemplate ValidatableTemplate, template templates.Reader) (*unstructured.Unstructured, error) {
-	i.completeLabels(*workload, apiTemplate.GetName(), apiTemplate.GetObjectKind().GroupVersionKind().Kind)
-
-	var (
-		blueprintParams []v1alpha1.BlueprintParam
-		err             error
-	)
-
-	if i.BlueprintParams != nil {
-		blueprintParams, err = i.BlueprintParams.GetBlueprintParams()
-		if err != nil {
-			return nil, fmt.Errorf("get blueprint params failed: %w", err)
-		}
-	}
-
-	paramMerger := realizer.NewParamMerger([]v1alpha1.BlueprintParam{}, blueprintParams, workload.Spec.Params)
-	params := paramMerger.Merge(template)
-
-	templatingContext, err := i.createTemplatingContext(*workload, params)
-	if err != nil {
-		return nil, fmt.Errorf("create templating context: %w", err)
-	}
-
-	stampContext := templates.StamperBuilder(workload, templatingContext, i.labels)
-	actualStampedObject, err := stampContext.Stamp(ctx, template.GetResourceTemplate())
-	if err != nil {
-		return nil, fmt.Errorf("could not stamp: %w", err)
-	}
-
-	return actualStampedObject, nil
-}
-
 var yttNotFound = errors.New("ytt must be installed in PATH but was not found")
 
 func ensureYTTAvailable(ctx context.Context) error {
@@ -296,15 +181,6 @@ func (i *TemplateFile) preprocessYtt(ctx context.Context) (string, error) {
 	}
 
 	return f.Name(), nil
-}
-
-func (i *TemplateTestGivens) completeLabels(workload v1alpha1.Workload, name string, kind string) {
-	i.labels = map[string]string{}
-
-	i.labels["carto.run/workload-name"] = workload.GetName()
-	i.labels["carto.run/workload-namespace"] = workload.GetNamespace()
-	i.labels["carto.run/template-kind"] = kind
-	i.labels["carto.run/cluster-template-name"] = name
 }
 
 func (i *TemplateTestGivens) createTemplatingContext(workload v1alpha1.Workload, params map[string]apiextensionsv1.JSON) (map[string]interface{}, error) {
@@ -370,52 +246,4 @@ func (i *TemplateTestGivens) getBlueprintInputs() (*Inputs, error) {
 	}
 
 	return &inputs, nil
-}
-
-// StringParam is a helper struct for use with the BuildBlueprintStringParams method
-// Either a Value or a DefaultValue should be specified for every StringParam
-// A Name is required for every StringParam
-type StringParam struct {
-	Name         string
-	Value        string
-	DefaultValue string
-}
-
-// BuildBlueprintStringParams is a helper method for creating string BlueprintParams for Givens.
-// BlueprintParams that hold other valid JSON values must be constructed by hand.
-func BuildBlueprintStringParams(candidateParams []StringParam) ([]v1alpha1.BlueprintParam, error) {
-	var completeParams []v1alpha1.BlueprintParam
-
-	for _, stringParam := range candidateParams {
-		newParam, err := buildBlueprintStringParam(stringParam.Name, stringParam.Value, stringParam.DefaultValue)
-		if err != nil {
-			return nil, fmt.Errorf("failed to build param: %w", err)
-		}
-		completeParams = append(completeParams, *newParam)
-	}
-
-	return completeParams, nil
-}
-
-func buildBlueprintStringParam(name string, value string, defaultValue string) (*v1alpha1.BlueprintParam, error) {
-	if (value == "" && defaultValue == "") ||
-		value != "" && defaultValue != "" {
-		return nil, fmt.Errorf("exactly one of value or defaultValue must be set")
-	}
-
-	if name == "" {
-		return nil, fmt.Errorf("name must be set")
-	}
-
-	param := v1alpha1.BlueprintParam{
-		Name: name,
-	}
-
-	if value != "" {
-		param.Value = &apiextensionsv1.JSON{Raw: []byte(fmt.Sprintf("%#v", value))}
-	} else {
-		param.DefaultValue = &apiextensionsv1.JSON{Raw: []byte(fmt.Sprintf("%#v", defaultValue))}
-	}
-
-	return &param, nil
 }
